feat(svelte5): data table
