+++ src/lib/index.ts
@@ -1,7 +1,8 @@
 import './scss/style.scss';
+
 export { default as Alert } from './Alert.svelte';
 export { default as Button } from './Button.svelte';
-// export { default as DataTable } from './DataTable.svelte';
+export { default as DataTable } from './DataTable.svelte';
 export { default as IconBurger } from './IconBurger.svelte';
 export { default as IconClose } from './IconClose.svelte';
 export { default as IconFloppyDisk } from './IconFloppyDisk.svelte';
@@ -16,7 +17,7 @@ export { default as NavBar } from './NavBar.svelte';
 export { default as Navigation } from './Navigation.svelte';
 export { default as NumberInput } from './NumberInput.svelte';
 export { default as RadioButton } from './RadioButton.svelte';
-export { default as TextInput } from './TextInput.svelte';
 export { default as SlottedComponentWrapper } from './SlottedComponentWrapper.svelte';
+export { default as TextInput } from './TextInput.svelte';
 
 export * from './definition';
